+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - You can give markdown descriptions to the metaskills to be displayed in the hovers by putting comments
   - You can prevent a file from being parsed by writing `# mythicscribe-disable file-parsing` as the first line
 - VSCode-Web Compatibility
+- `MythicScribe.showNodeGraph` command! Shows a graph of all the configurations parsed. Can be filtered based on configuration type and whether or not it's an open editor
 - Color Picker for the relevant attributes and options.
 - Basic support for completions and hovers for conditions inside of certain mechanic attributes, such as projectile's hitconditions
 #### Configurations
@@ -21,10 +22,10 @@
 - "Bundle" Custom Datasets
   - A Bundle json can contain references to any number of other custom dataset types
   - When a Bundle is imported as a File and not as a Link, the path of other File-sourced datasets is interpreted as "relative" to the Bundle's location
-#### Custom Datasets
+#### Datasets
 - `Remove Custom Dataset` command to remove a Custom Dataset without having to edit the settings file
 - `Create Bundle Dataset` command to create a Bundle Dataset based on other Custom Datasets you have previously added
-- `Load Datasets` command to reload the datasets manually
+- `Load Datasets` command to reload all the datasets (custom and not) manually
 - `Open Logs` command to open the Extension's logs. For debugging purposes
 
 ### Changed